move to pure sql. pass all existing tests

+++ lmwrapper/caching.py
@@ -60,4 +60,5 @@ def get_disk_cache():
 
 def clear_cache_dir():
     import shutil
-    shutil.rmtree(cache_dir())
+    if cache_dir().exists():
+        shutil.rmtree(cache_dir())
